LPD-92619 Export scoped configurations with the instance
The headless export endpoint copies the instance's scoped OSGi configurations into the exported schema, but the portlet action did not, so an export triggered from the Virtual Instances UI produced a schema missing them. The step is a no-op when database partitioning is enabled, because the exported partition already carries the configurations.

Rather than add a third copy of that logic, it moves to a PortalInstanceConfigurationExporter component in portal-instances-service, published through a new portal-instances-api module so every caller can reach it. The headless endpoint and ExportPortalInstanceOperation now delegate to it instead of carrying their own copies, which had already drifted: the copy in ExportPortalInstanceOperation dereferenced the configuration dictionary without a null check, so a Configuration_ row with a null dictionary threw a NullPointerException there while the other copy skipped it. The shared component keeps the null check, so that failure is fixed for every caller.

Neither this commit nor the test commit above it is required by the story's acceptance criteria, and the two can be dropped together.
