+++ modules/apps/portal-instances/portal-instances-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Instances API
+Bundle-SymbolicName: com.liferay.portal.instances.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.instances.configuration
