+++ modules/apps/portal-instances/portal-instances-api/src/main/resources/com/liferay/portal/instances/configuration/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
